+++ src/Recyclarr.Core/Config/Parsing/PostProcessing/ConfigMerging/ConfigIncludeProcessor.cs
@@ -19,7 +19,10 @@ public class ConfigIncludeProcessor(IFileSystem fs, IAppPaths paths, ILogger log
         var configFile = ConvertToAbsolute(include.Config);
         if (configFile?.Exists != true)
         {
-            throw new YamlIncludeException($"Include path could not be resolved: {include.Config}");
+            throw new YamlIncludeException(
+                $"Include path could not be resolved: {include.Config}. "
+                    + "Relative paths must exist in the `includes` directory."
+            );
         }
 
         return configFile;
@@ -37,19 +40,6 @@ public class ConfigIncludeProcessor(IFileSystem fs, IAppPaths paths, ILogger log
         if (fsPath.Exists)
         {
             log.Debug("Path rooted to the includes directory: {Path}", path);
-            return fsPath;
-        }
-
-        fsPath = paths.ConfigsDirectory.File(path);
-        // ReSharper disable once InvertIf
-        if (fsPath.Exists)
-        {
-            log.Warning(
-                "DEPRECATED: Include templates inside the `configs` directory are no longer supported. "
-                    + "These files should be relocated to the new sibling `includes` directory instead. "
-                    + "See: <https://recyclarr.dev/guide/upgrade-guide/v8.0/#include-dir>"
-            );
-
             return fsPath;
         }